requirements.txt
requests==2.28.2
matplotlib==3.8.1
cartopy==0.21.0
gunicorn==20.1.0
basemap==1.4.1
Flask==2.2.3
xarray==2023.6.0
pillow==9.5.0
Werkzeug==2.2.2
schedule==1.1.0